+++ cmd/jl/errors.go
@@ -19,7 +19,4 @@ package main
 
 import "errors"
 
-var (
-	ErrForbiddenTemplateAndColumnFlags = errors.New("using both flags template and columns is forbidden")
-	ErrInvalidRawType                  = errors.New("invalid raw type")
-)
+var ErrInvalidRawType = errors.New("invalid raw type")